feat(api)!: Authentication and User Check
When a user logs into the system, their details are returned (displayed) on the nav-user component. "Log out" has been added in order
to deal with tokens efficiently by removing them when not in use and for a user to also log out of the system when its not in use.

Additions & Fixes:
- fix: Fix user type issue when checking for authenticated user and returning details in nav-user component
- Add TimeOff logo and icon
- build: Fix Project Build by moving endpoints/api file to convenient folder
- Remove clutter (unwanted files)
- style: Update ThemeProvider to use "dark" instead of "light"
- Add logoutUser endpoint to provide logout functionality