update: Limit site access through IP address restriction
As requested, the site can only be accessed when the user is connected to specified networks (noted by ALLOWED_IPS).
Middleware has been created to handle the restriction on both the frontend and backend of the system. The first commit
pushes the frontend middleware and the backend middleware will follow